% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.pred.pres.R
\name{a.pred.pres}
\alias{a.pred.pres}
\title{Calculation of areas predicted under each value of suitability}
\usage{
a.pred.pres(inrast)
}
\arguments{
\item{inrast}{a RasterLayer of the ecological niche model to be evaluated, but standardized
previously.}
}
\description{
a.pred.pres calculates proportionate areas predicted under each value of suitability.
This function is designed to be used specifically in the \code{\link{ku.enm.proc}} function.
}
