% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_pred_pres.R
\name{a_pred_pres}
\alias{a_pred_pres}
\title{Calculation of areas predicted under each value of suitability}
\usage{
a_pred_pres(inrast)
}
\arguments{
\item{inrast}{a RasterLayer of the ecological niche model to be evaluated, but standardized
previously.}
}
\description{
a_pred_pres calculates proportionate areas predicted under each value of suitability.
This function is designed to be used specifically in the \code{\link{kuenm_proc}} function.
}
