% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_analysis_helpers.R
\name{aicc}
\alias{aicc}
\title{AICc calculation of Maxent SWD predictions}
\usage{
aicc(occ, prediction, npar)
}
\arguments{
\item{occ}{numerical matrix containing coordinates of the occurrences used to
create the ecological niche models to be evaluated; columns must be: longitude
and latitude.}

\item{prediction}{matrix of longitude and latidue coordinates, and Maxent Raw
predictions obtained using the SWD format. Prediction coordinates must include
the ones in \code{occ}}

\item{npar}{(numeric) number of parameters of the model. Use function \code{\link{n.par}}
to obtain number of parameters for each model from lambdas file.}
}
\value{
A data.frame with results from AICc analyses.
}
\description{
aicc calculates the Akaike information criterion corrected for
small sample sizes (AICc) to single or multiple models produced with Maxent.
}
\details{
This function is a modification of calc.aicc from the ENMeval package to allow
calculations using numerical values instead raster predictions.
}
