% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_swd.R
\name{all_var_comb}
\alias{all_var_comb}
\title{Helper to create all variable combinations}
\usage{
all_var_comb(var.names, min.number = 2)
}
\arguments{
\item{var.names}{(character) vector of variable names}

\item{min.number}{(numeric) minimum number of variables per set.}
}
\value{
A list of vectors containing variable names per set.
}
\description{
Helper to create all variable combinations
}
