% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_functions.R
\name{auc_comp}
\alias{auc_comp}
\title{Helper function to compute partial AUC, AUC ratio.}
\usage{
auc_comp(iterationno, occurtbl, rand.percent, omissionval, classpixels)
}
\arguments{
\item{iterationno}{number of iteration to compute partial AUC values}

\item{occurtbl}{Validation data. Must have 3 columns SpName, Longitude, Latitude.}

\item{rand.percent}{Occurrence points to be sampled randomly from the test data for bootstrapping.}

\item{omissionval}{1-E.}

\item{classpixels}{Pixel classes.}
}
\description{
Helper function to compute partial AUC, AUC ratio.
}
