% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.auc.R
\name{calc.auc}
\alias{calc.auc}
\title{Calculation of partial areas under the ROC curve}
\usage{
calc.auc(xytable, omissionval, iterationno)
}
\arguments{
\item{xytable}{a XY coordinates table prepared with the \code{\link{g.xy.tab}} function.}

\item{omissionval}{(numeric) value from 0 to 100 that will be used as threshold (E).}

\item{iterationno}{(numeric) number of bootstrap iterations to be performed.}
}
\description{
calc.auc calculates the area under the curve using a threshold (E) value.
This function is designed to be used specifically in the \code{\link{ku.enm.proc}} function.
}
