% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculation of partial areas under the ROC curve}
\usage{
calc_auc(xytable, omissionval, iterationno)
}
\arguments{
\item{xytable}{a XY coordinates table prepared with the \code{\link{g_xy_tab}} function.}

\item{omissionval}{(numeric) value from 0 to 100 that will be used as threshold (E).}

\item{iterationno}{(numeric) number of bootstrap iterations to be performed.}
}
\description{
calc_auc calculates the area under the curve using a threshold (E) value.
This function is designed to be used specifically in the \code{\link{kuenm_proc}} function.
}
