% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_var_contrib.R
\name{explore_var_contrib}
\alias{explore_var_contrib}
\alias{plot_contribution}
\title{Evaluation and plot of variable contribution to single Maxent models}
\usage{
explore_var_contrib(occ, M_variables, maxent.path, reg.mult = 1,
                    f.clas = NULL, max.memory = 1000, args = NULL,
                    plot = TRUE)

plot_contribution(contribution_list, col.cont = "gray25",
                  col.imp = "gray25", col.with = "gray15",
                  col.without = "gray65", col.all = "gray1")
}
\arguments{
\item{occ}{a data.frame with occurrence records. If \code{M_variables} is of
class "RasterStack", columns must be (in that order): Species, Longitude,
Latitude. If \code{M_variables} is of class "data.frame" (SWD), columns
containing the values of the environmental variables in \code{M_variables}
must be added.}

\item{M_variables}{RasterStack or data.frame. If RasterStack, object containing
the variables to be used for modeling. If data.frame (SWD), columns must be:
"background", "longitude", "latitude", and a column for each of the variables
of interest. If data.frame, \code{occ} must be a data.frame.}

\item{maxent.path}{(character) the path were maxent.jar file is in your computer.}

\item{reg.mult}{(numeric vector) regularization multiplier(s) to be evaluated.}

\item{f.clas}{(character) feature classes can be selected from five different
combination sets or manually. Combination sets are: "all", "basic", "no.t.h",
"no.h", and "no.t". Default = "all". basic = "l", "lq", "lqp", "lqpt", "lqpth".
Combinations "no.t.h", "no.h", and "no.t", exclude t and/or h. See details for
all the available potential combinations of feature classes.}

\item{max.memory}{(numeric) maximum memory (in megabytes) to be used by Maxent
while creating the models. Default = 1000.}

\item{args}{(character) additional arguments that can be passed to Maxent.
See the Maxent help for more information on how to write these arguments,
default = NULL. Note that some arguments cannot be changed here because they
are part of the parameters of the function already.
See details for other options.}

\item{sample.size}{(numeric) number of points to represent the background for
the model. Default = 10000}

\item{plot}{(logical) whether to plot results.}

\item{contribution_list}{a list of results obtained with
\code{\link{explore_var_contrib}}, or an element of a list obtained with
\code{\link{model_var_contrib}}.}

\item{col.cont}{color of contribution bars; default = "gray25".}

\item{col.imp}{color of importance bars; default = "gray25".}

\item{col.with}{color of regularized training gain bars when using individual
variables; default = "gray15".}

\item{col.without}{color of regularized training gain bars when using all
variables except the one in question; default = "gray65".}

\item{col.all}{color of line representing regularized training gain of the
model using all variables; default = "gray1".}
}
\value{
A list with results of variable contribution, permutation importance, and
jackknife results. If \code{plot} = TRUE results are plotted as horizontal
bars using \code{\link{plot_contribution}}.
}
\description{
explore_var_contrib helps to explore variable contribution to
single maxent models based on metrics of contribution percentage, permutation
importance, and a jackknife analysis.
}
\details{
All potential combinations of feature classes
(l = linear, q = quadratic, p = product, t = threshold, and h = hinge) are:
"l", "q", "p", "t", "h", "lq", "lp", "lt", "lh", "qp", "qt", "qh", "pt", "ph",
"th", "lqp", "lqt", "lqh", "lpt", "lph", "lth", "qpt", "qph", "qth", "pth",
"lqpt", "lqph", "lqth", "lpth", "qpth", and "lqpth".

The way to include further arguments is as follows:
args = "biasfile=COMPLETE_PATH\\bias.asc biastype=3" in windows,
or args = "biasfile=COMPLETE_PATH/bias.asc biastype=3" in Unix based systems.
If the path contains spaces the way to write it is:
args = "biasfile=\"COMPLETE PATH\\bias.asc\" biastype=3" in windows, or
args = "biasfile=\"COMPLETE PATH/bias.asc\" biastype=3" in Unix based systems.

Other options that can be included in args are all "Flags" from the following
list:

Flag | Abbrv | Type | Default | Meaning
\itemize{
\item maximumbackground | MB | integer | 10000 | If the number of background points / grid cells is larger than this number, then this number of cells is chosen randomly for background points.
\item togglelayertype | t | string | | Toggle continuous/categorical for environmental layers whose names begin with this prefix (default: all continuous).
\item biasfile | | file | | Sampling is assumed to be biased according to the sampling distribution given in this grid file. Values in this file must not be zero or negative. MaxEnt will factor out the bias. We recomend to create this file as a kernell density of geographic points representing all localities were samplings of similar organisms have been performed (multiply this layer by 1000 and round it to reduce number of decimals). IMPORTANT: A biasfile must be included with its entire path, as indicated above above.
\item biastype | | integer | | If biasfile is defined, this integer needs to be definef depending on the type of bias added. If the bias file is prepared as recomended, biastype=3.
\item writebackgroundpredictions | | boolean | FALSE | Write .csv file with predictions at background points.
\item maximumiterations | m | integer | 500 | Stop training after this many iterations of the optimization algorithm.
\item convergencethreshold | c | double | 0.00001 | Stop training when the drop in log loss per iteration drops below this number.
\item threads | | integer | 1 | Number of processor threads to use. Matching this number to the number of cores on your computer speeds up some operations, especially variable jackknifing.
\item logfile | | string | maxent.log | File name to be used for writing debugging information about a run in output directory.
\item cache | | boolean | TRUE | Make a .mxe cached version of ascii files, for faster access.
\item defaultprevalence | | double | 0.5 | Default prevalence of the species: probability of presence at ordinary occurrence points. See Elith et al., Diversity and Distributions, 2011 for details.
}

Other more advanced arguments are (use these ones only if you understand them completely):
\itemize{
\item lq2lqptthreshold | | integer | 80 | Number of samples at which product and threshold features start being used.
\item l2lqthreshold | | integer | 10 | Number of samples at which quadratic features start being used.
\item hingethreshold | | integer | 15 | Number of samples at which hinge features start being used.
\item beta_threshold | | double | -1 | Regularization parameter to be applied to all threshold features; negative value enables automatic setting.
\item beta_categorical | | double | -1 | Regularization parameter to be applied to all categorical features; negative value enables automatic setting.
\item beta_lqp | | double | -1 | Regularization parameter to be applied to all linear, quadratic and product features; negative value enables automatic setting.
\item beta_hinge | | double | -1 | Regularization parameter to be applied to all hinge features; negative value enables automatic setting.
}
}
\examples{
# data
occ <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                           pattern = "sp_joint.csv", full.names = TRUE))
occ <- data.frame(Species = "A_americanum", occ)

mvars <- raster::stack(list.files(system.file("extdata", package = "kuenm"),
                                  pattern = "Mbio_", full.names = TRUE))
# analysis
var_cont <- explore_var_contrib(occ = occ, M_variables = mvars,
                                maxent.path = "C:/Maxent/3.4.1", plot = FALSE)

# plot
plot_contribution(var_cont)
}
