% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{ext_type}
\alias{ext_type}
\title{Helper function to select extrapolation options}
\usage{
ext_type(ext.type = "all")
}
\arguments{
\item{ext.type}{(character) extrapolation type to be used. Options are:
"all", "ext_clam", "ext", and "no_ext". Default = "all".}
}
\description{
Helper function to select extrapolation options
}
