% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{feature_classes}
\alias{feature_classes}
\title{Helper function to select feature classes}
\usage{
feature_classes(f.clas = "all")
}
\arguments{
\item{f.clas}{(character) feature clases can be selected from five different
combination sets or manually. Combination sets are: "all", "basic", "no.t.h",
"no.h", and "no.t". Default = "all". basic = "l", "lq", "lqp", "lqpt", "lqpth".
Combinations "no.t.h", "no.h", and "no.t", exclude t and/or h. See details for
all the available potential combinations of feature classes.}
}
\description{
Helper function to select feature classes
}
\details{
Below all potential combinations of feature classes are shown. Manual selection
can be done by creating a vector of one or more of the combinations of this
list. l = linear, q = quadratic, p = product, t = threshold, and h = hinge.
"l", "q", "p", "t", "h", "lq", "lp", "lt", "lh", "qp", "qt", "qh", "pt", "ph",
"th", "lqp", "lqt", "lqh", "lpt", "lph", "qpt", "qph", "qth", "pth", "lqpt",
"lqph", "lqth", "lpth", and "lqpth".
}
