% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.xy.tab.R
\name{g.xy.tab}
\alias{g.xy.tab}
\title{Preparing an XY coordinates table for calculation of the partial ROC}
\usage{
g.xy.tab(classpixels, occurinclass)
}
\arguments{
\item{classpixels}{table with information about proportionate areas predicted under each
value of suitability.}

\item{occurinclass}{table with the percentage of points within each class of suitability.}
}
\description{
g.xy.tab is a function designed to be used specifically in the
\code{\link{ku.enm.proc}} function. It generates an XY coordinates table to
calculate partial areas under the ROC curve.
}
