% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_xy_tab.R
\name{g_xy_tab}
\alias{g_xy_tab}
\title{Preparing an XY coordinates table for calculation of the partial ROC}
\usage{
g_xy_tab(classpixels, occurinclass)
}
\arguments{
\item{classpixels}{table with information about proportionate areas predicted under each
value of suitability.}

\item{occurinclass}{table with the percentage of points within each class of suitability.}
}
\description{
g_xy_tab is a function designed to be used specifically in the
\code{\link{kuenm_proc}} function. It generates an XY coordinates table to
calculate partial areas under the ROC curve.
}
