% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_functions.R
\name{generatexytableb}
\alias{generatexytableb}
\title{Helper function to compute the area (number of pixels) that a certain threshold has.}
\usage{
generatexytableb(classpixels, occurinclass)
}
\arguments{
\item{classpixels}{Pixel threshold class .}

\item{occurinclass}{Ocurrence points that lies in a certain class.}
}
\description{
Helper function to compute the area (number of pixels) that a certain threshold has.
}
