% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.free.ram.R
\name{get.free.ram}
\alias{get.free.ram}
\title{Getting available RAM}
\usage{
get.free.ram()
}
\value{
A numeric value representing the amount of RAM that is not being used in other processes.
}
\description{
get.free.ram is used by the \code{\link{ku.enm.cal}} and \code{\link{ku.enm.mod}} functions.
}
\details{
Although this function was designed to be used with the \code{\link{ku.enm.cal}} and
\code{\link{ku.enm.mod}} functions, it may be used to know how much RAM is free in a specific time.
}
\examples{
fre_ram <- get.free.ram()
}
