% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_free_ram.R
\name{get_free_ram}
\alias{get_free_ram}
\title{Getting available RAM}
\usage{
get_free_ram()
}
\value{
A numeric value representing the amount of RAM that is not being used in other processes.
}
\description{
get_free_ram is used by the \code{\link{kuenm_cal}} and \code{\link{kuenm_mod}} functions.
At the moment this functions works well on Windows, for MAC and Linux it returns 4 GB.
}
\details{
Although this function was designed to be used with the \code{\link{kuenm_cal}} and
\code{\link{kuenm_mod}} functions, it may be used to know how much RAM is free at any time.
}
\examples{
fre_ram <- get_free_ram()
}
