% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{gvars}
\alias{gvars}
\title{Variables masked to the area where a model is transferred}
\format{
A RasterStack with 900 rows, 2160 columns, 1944000 cells, and 4 layers:
\describe{
\item{Temperature}{temperature, in Celsius degrees times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A RasterStack containing predictor variables masked to the area where a model
is projected. Variables represent four future bioclimatic variables (2050) of
the NCAR-CCSM4 general circulation model under the RCP 8.5 emission scenario.
}
\examples{
gvars <- raster::stack(list.files(system.file("extdata", package = "kuenm"),
                                  pattern = "Gbio_", full.names = TRUE))

summary(gvars)
}
