% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvars_mop.R
\docType{data}
\name{gvars_mop}
\alias{gvars_mop}
\title{Variables of the area to which the ecological niche model is transferred.}
\format{A RasterStack with 900 rows, 2160 columns, 1944000 cells, and 4 layers:
\describe{
  \item{Temperature}{tempreature, in celcius degrees times 10.}
  \item{Precipitation}{precipitation, in milimiters.}
  ...
}}
\source{
\url{http://www.ku***/}
}
\usage{
gvars_mop
}
\description{
A dataset containing predictor variables of the area of projection of the ecological
niche model. Variables represent four future bioclimatic variables (2050) of the
NCAR-CCSM4 Global Circulation Model in the RCP 8.5 emission scenario.
}
\keyword{datasets}
