% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierpart_analyses.R
\name{hierpart_analyses}
\alias{hierpart_analyses}
\title{Performing hierarchical partitioning analyses}
\usage{
hierpart_analyses(tables.folder, out.dir, kept = FALSE, verbose = TRUE)
}
\arguments{
\item{tables.folder}{(character) name of the folder where the tables created
with the function {\link{hierpart_tables}} are.}

\item{out.dir}{(character) name of the output directory where results will
be written.}

\item{kept}{(logical) whether to keep tables in \code{tables.folder}.
Default = FALSE.}

\item{verbose}{(logical) whether to print messages; default = TRUE.}
}
\value{
A data.frame containing the summary of total effects of factors on the model.

All other results will be written in \code{out.dir}.
}
\description{
Helper function to perform hierarchical partitioning analysis
of the variance coming from distinct sources in ecological niche models.
}
