% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierpart_tables.R
\name{hierpart_tables}
\alias{hierpart_tables}
\title{Prepare tables for hierarchical partitioning}
\usage{
hierpart_tables(model.names, sp.name, replicate.numbers, format = "asc",
  parameters, clim.models, emi.scenarios, iterations = 100,
  sample.size = 100, out.dir)
}
\arguments{
\item{model.names}{(character) vector of model names.}

\item{sp.name}{(character) species names as it is in the csv file used for creating
the model.}

\item{replicate.numbers}{(numeric) vector of replicate numbers.}

\item{format}{(character) format of model raster files. Options are: "asc" or "tif"; default = "asc".}

\item{parameters}{(character) vector of parameter settings names.}

\item{clim.models}{(character) vector of climatic model names.}

\item{emi.scenarios}{(character) vector of emission scenario names.}

\item{iterations}{(numeric) number of iterations to be performed; default = 100.}

\item{sample.size}{(numeric) number of pixels to be sampled per each model. Default = 100.}

\item{out.dir}{(character) name of the output directory where matrices will be written.}
}
\value{
All the tables needed to performing hierarchical partitioning analyses of the
varianze in ecological niche models.
}
\description{
Helper function to prepare tables for hierarchical partitioning analyses
of the varianze comming from distinct sources in ecological niche models.
}
\details{
At least two of the following sources of variation must exist to perform
the analysis: \code{replicate.numbers}, \code{parameters}, \code{clim.models},
or \code{emi.scenarios}.
}
