% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.eval.R
\name{html.eval}
\alias{html.eval}
\title{Creation of an html file with results from model calibration}
\usage{
html.eval(path = getwd(), file.name)
}
\arguments{
\item{path}{directory in wich the html file will be written, default = current directory.}

\item{file.name}{(character) name of the html file.}
}
\value{
An html file summarizing all results from the model calibration and selection process.
}
\description{
html.eval creates an html file that summarizes all outputs from
the model calibration and selection processes.
}
\details{
This function is used along with the \code{\link{ku.enm.ceval}} function.
}
\examples{
path <- getwd()
name <- "evaluation_results"

html_file <- html.eval(path = path, file.name = name)
}
