% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_calibration.R
\name{html_calibration}
\alias{html_calibration}
\title{Creation of an HTML file with results from model calibration}
\usage{
html_calibration(path = getwd(), file.name)
}
\arguments{
\item{path}{directory where the HTML file will be written; current directory
by default.}

\item{file.name}{(character) name of the HTML file without extension (e.g.,
"calibration_results")}
}
\value{
An HTML file summarizing results from model calibration, evaluation, and
selection.
}
\description{
html_calibration creates an HTML file that summarizes all outputs
from model calibration, evaluation, and selection.
}
\details{
This function is used along with the functions \code{\link{kuenm_ceval}}
\code{\link{kuenm_cal_swd}}.
}
\examples{
path <- getwd() # directory with outputs of the kuenm_ceval function
name <- "evaluation_results"

\dontrun{
html_calibration(path = path, file.name = name)
}
}
