% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_eval.R
\name{html_eval}
\alias{html_eval}
\title{Creation of an HTML file with results from model calibration}
\usage{
html_eval(path = getwd(), file.name)
}
\arguments{
\item{path}{directory in which the HTML file will be written; default = current directory.}

\item{file.name}{(character) name of the HTML file.}
}
\value{
An HTML file summarizing results from the model calibration and selection process.
}
\description{
html_eval creates an HTML file that summarizes all outputs from
the model calibration and selection processes.
}
\details{
This function is used along with the \code{\link{kuenm_ceval}} function.
}
\examples{
path <- getwd() # directory with outputs of the kuenm_ceval function
name <- "evaluation_results"

html_file <- html_eval(path = path, file.name = name)
}
