% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.ceval.R
\name{ku.enm.ceval}
\alias{ku.enm.ceval}
\title{Evaluation of candidate Maxent models during calibration}
\usage{
ku.enm.ceval(path, occ.joint, occ.tra, occ.test, batch, out.eval,
  threshold = 5, rand.percent = 50, iterations = 500, kept = TRUE,
  selection = "OR_AICc")
}
\arguments{
\item{path}{(character) directory in wich folders containig calibration models are being or
were created.}

\item{occ.joint}{(character) is the  csv file with training and testing occurrences combined,
columns must be: species, longitud, latitud.}

\item{occ.tra}{(character) is the name of the csv file with the training occurrences,
columns equal to occ.joint.}

\item{occ.test}{(character) is the name of the csv file with the evaluation occurrences,
columns equal to occ.joint.}

\item{batch}{(character) name of the batch file with the code to create all candidate Maxent models
for calibration.}

\item{out.eval}{(character) name of the folder were evaluation results will be written.}

\item{threshold}{(numeric) is the percentage of omission error allowed, default = 5.}

\item{rand.percent}{(numeric) is the percentage of data to be used for the bootstraping process
when calculating partial ROCs, default = 50.}

\item{iterations}{(numeric) is the number of times that the bootstrap is going to be repeated,
default = 500.}

\item{kept}{(logical) if false all candidate models will be erased after evaluation, default = TRUE.}

\item{selection}{(character) model selection criterion, can be "OR_AICc", "AICc", or "OR";
OR = omission rates. Default = "OR_AICc", which means that among statistically significant models,
those with omission rates below the threshold and among them those with delta AICc up to 2 will be
presented. If "AICc" criterion is chosen, significant models with delta AICc up to 2 will be presented.
If "OR" is chosen, the 10 first significant models with the lowest omission rates will be presented.}
}
\value{
A folder, in the working directory, containing: a csv file with information about models meeting
the user-defined selection criterion; another csv file with a summary of the evaluation and selection
process; an extra csv file containing all the statistics of model performance (pROC, AICc, and
omission rates) for all candidate models; a png scatterplot of all models based on the AICc values and
omission rates; and, an html file sumarizing all the information produced after evaluation for helping with
further interpretations.
}
\description{
ku.enm.ceval evaluates candidate Maxent models in terms of statistical
significance (partial ROC), prediction ability (omission rates), and complexity (AICc).
After evaluation this function selects the very best models based on a user-defined
criterion.
}
\details{
This function is used after or during the creation of Maxent candidate models for
calibration.
}
