% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.feval.R
\name{ku.enm.feval}
\alias{ku.enm.feval}
\title{Evaluation of final Maxent models with independent data}
\usage{
ku.enm.feval(path, occ.joint, occ.ind, replicates, out.eval, threshold = 5,
  rand.percent = 50, iterations = 500)
}
\arguments{
\item{path}{(character) directory in wich folders containig final models are being or
were created.}

\item{occ.joint}{(character) is the  csv file with training and testing occurrences combined,
or the file containing occurrences used to create final models, columns must be: species,
longitud, latitud.}

\item{occ.ind}{(character) is the name of the csv file with independent occurrences for model
evaluation, this occurrences were not used when creating final models, columns equal to occ.joint.}

\item{replicates}{(logical) whether or not final models where created performing replicates.}

\item{out.eval}{(character) name of the folder were evaluation results will be written.}

\item{threshold}{(numeric) is the percentage of omission error allowed, default = 5.}

\item{rand.percent}{(numeric) is the percentage of data to be used for the bootstraping process
when calculating partial ROCs, default = 50.}

\item{iterations}{(numeric) is the number of times that the bootstrap is going to be repeated,
default = 500.}
}
\value{
A folder, in the working directory, containing a csv file with the final models evaluation
results.
}
\description{
ku.enm.feval evaluates final Maxent models in terms of statistical
significance (partial ROC) and omission rates with a user-defined threshold.
}
\details{
This function is used after finishing the creation of final Maxent models.
}
