% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.mop.R
\name{ku.enm.mop}
\alias{ku.enm.mop}
\title{Extrapolation risk analysis for single comparisons}
\usage{
ku.enm.mop(M.stack, G.stack, percent = 10, normalized = TRUE)
}
\arguments{
\item{M.stack}{a raster stack of variables representing the M area.}

\item{G.stack}{a raster stack of variables representing the G area, areas or scenarios to
which models are transferred.}

\item{percent}{(numeric) percetage of values sampled from M to calculate the MOP.}

\item{normalized}{(logical) if true values of similarity are presented from 0 to 1,
default = TRUE.}
}
\value{
A Mobility-Oriented Parity raster layer.
}
\description{
ku.enm.mop calculates a Mobility-Oriented Parity layer by
comparing environmental values between the M area and the area or
scenario to which an ecological niche model is transferred.
}
\details{
The MOP is calculated following Owens et al., 2013
\url{https://doi.org/10.1016/j.ecolmodel.2013.04.011}.
}
\examples{
data()
mvars <- mvars_mop
gvars <- gvars_mop
perc <- 10
norm <- TRUE

mop <- ku.enm.mop(M.stack = mvars, G.stack = gvars,
                  percent = perc, normalized = norm)
}
