% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.omrat.R
\name{ku.enm.omrat}
\alias{ku.enm.omrat}
\title{Omission rates calculation for single models}
\usage{
ku.enm.omrat(model, threshold = 5, occ.tra, occ.test)
}
\arguments{
\item{model}{a raster layer of the model to be evaluated.}

\item{threshold}{(numeric) value(s) from 0 to 100 that will be used as thresholds,
default = 5.}

\item{occ.tra}{a numeric matrix containing longitude and latitude of the
occurrences used to create the ecological niche model to be evaluated.}

\item{occ.test}{a numeric matrix containing longitude and latitude of the
occurrences used to test the ecological niche model to be evaluated.}
}
\value{
A named numeric value or numeric vector with the result(s).
}
\description{
ku.enm.omrat calculates omission rates of geographic projections
of ecologicla niche models based in one or multiple user specified threshold.
}
\examples{
data()

#single threshold
model <- sp_mod
thres <- 5
octr <- sp_train
octe <- sp_test

om_rate <- ku.enm.omrat(model, threshold = thres,
                        occ.tra = octr, occ.test = octe)

#multiple thresholds
thres1 <- c(5, 10, 20)

om_rate <- ku.enm.omrat(model, threshold = thres1,
                        occ.tra = octr, occ.test = octe)
}
