% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.proc.R
\name{ku.enm.proc}
\alias{ku.enm.proc}
\title{Partial ROC calculation of single models}
\usage{
ku.enm.proc(occ.test, model, threshold = 5, rand.percent = 50,
  iterations = 1000)
}
\arguments{
\item{occ.test}{a numeric matrix containing longitude and latitude of the
occurrences used to test the ecological niche model.}

\item{model}{a raster layer of the ecological niche model to be evaluated.}

\item{threshold}{(numeric) value from 0 to 100 that will be used as threshold,
default = 5.}

\item{rand.percent}{(numeric) value from 0 to 100 representing the amount of data
to be used for performing the bootstrap process for calculating the partial ROC,
default = 50.}

\item{iterations}{(numeric) number of bootstrapped iterations to be performed,
default = 1000.}
}
\value{
A list containing a named vector with the final partial ROC results and
a matrix containing the AUC values and AUC ratios calculated for each iteration.
}
\description{
ku.enm.proc calculates the partial ROC of single models.
}
\details{
Partial ROC is calculated following Peterson et al., 2008
\url{https://doi.org/10.1016/j.ecolmodel.2007.11.008}.
}
\examples{
data()
occ <- sp_test
model <- sp_mod
thres <- 5
rand_perc <- 50
iterac <- 100

p_roc <- ku.enm.proc(occ.test = occ, model = model, threshold = thres,
                   rand.percent = rand_perc, iterations = iterac)
}
