% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_aicc.R
\name{kuenm_aicc}
\alias{kuenm_aicc}
\title{AICc calculation for Maxent models}
\usage{
kuenm_aicc(occ, model, npar)
}
\arguments{
\item{occ}{matrix or data.frame with coordinates of the occurrences used to
create the model (raster) to be evaluated; columns must be: longitude and
latitude.}

\item{model}{a RasterLayer or RasterStack of model projections created using
Maxent with raw outputs.}

\item{npar}{(numeric) number of parameters for \code{model}. Length must
correspond with number of layers in \code{model}. See function
\code{\link{n_par}}.}
}
\value{
A data.frame containing values of AICc, delta AICc, weight of AICc, and
number of parameters. The number of rows of the data.frame corresponds to
the number of models evaluated.
}
\description{
kuenm_aicc calculates the Akaike information criterion corrected
for small sample sizes (AICc) for single or multiple models produced with
Maxent.
}
\details{
Calculations are done following
\href{https://doi.org/10.1890/10-1171.1}{Warren and Seifert (2011)}.
}
\examples{
data("sp_joint", package = "kuenm")
model <- raster::raster(system.file("extdata/sp_model_joint.tif",
                                             package = "kuenm"))

lbds <- readLines(system.file("extdata/lambdas_model_joint.lambdas",
                              package = "kuenm"))
npar <- n_par(lbds) # counting number of parameters

aicc <- kuenm_aicc(occ = sp_joint, model = model, npar = npar)
}
\seealso{
\code{\link{aicc}} for results obtained using the SWD format.
}
