% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_ceval.R
\name{kuenm_ceval}
\alias{kuenm_ceval}
\title{Evaluation of candidate Maxent models during calibration}
\usage{
kuenm_ceval(path, occ.joint, occ.tra, occ.test, batch, out.eval,
  threshold = 5, rand.percent = 50, iterations = 500, kept = TRUE,
  selection = "OR_AICc")
}
\arguments{
\item{path}{(character) directory in which folders containig calibration models are being created
or were created.}

\item{occ.joint}{(character) the name of csv file with training and testing occurrences combined;
columns must be: species, longitude, latitude.}

\item{occ.tra}{(character) the name of the csv file with the training occurrences;
columns as in occ.joint.}

\item{occ.test}{(character) the name of the csv file with the evaluation occurrences;
columns as in occ.joint.}

\item{batch}{(character) name of the batch file (bash for Unix) with the code to create all candidate models
for calibration.}

\item{out.eval}{(character) name of the folder where evaluation results will be written.}

\item{threshold}{(numeric) the percentage of training data omission error allowed (E); default = 5.}

\item{rand.percent}{(numeric) the percentage of data to be used for the bootstraping process
when calculating partial ROCs; default = 50.}

\item{iterations}{(numeric) the number of times that the bootstrap is going to be repeated;
default = 500.}

\item{kept}{(logical) if FALSE, all candidate models will be erased after evaluation, default = TRUE.}

\item{selection}{(character) model selection criterion, can be "OR_AICc", "AICc", or "OR";
OR = omission rates. Default = "OR_AICc", which means that among models that are statistically significant
and that present omission rates below the threshold, those with delta AICc up to 2 will be
selected. If "AICc" criterion is chosen, all significant models with delta AICc up to 2 will be selected
If "OR" is chosen, the 10 first significant models with the lowest omission rates will be selected.}
}
\value{
A list with three dataframes containing results from the calibration process. In addition,
a folder, in the working directory, containing a csv file with information about models meeting
the user-defined selection criterion, another csv file with a summary of the evaluation and selection
process, an extra csv file containing all the statistics of model performance (pROC, AICc, and
omission rates) for all candidate models, a png scatterplot of all models based on the AICc values and
omission rates, and an HTML file sumarizing all the information produced after evaluation for helping with
further interpretation.
}
\description{
kuenm_ceval evaluates candidate models in terms of statistical
significance (partial ROC), prediction ability (omission rates), and model complexity (AICc).
After evaluation, this function selects the best models based on user-defined criteria.
}
\details{
This function is used after or during the creation of Maxent candidate models for calibration.
}
