% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_feval_swd.R
\name{kuenm_feval_swd}
\alias{kuenm_feval_swd}
\title{Evaluation of final Maxent models with independent data in SWD format}
\usage{
kuenm_feval_swd(path, occ.joint, occ.ind, replicates, out.eval, threshold = 5,
                rand.percent = 50, iterations = 500)
}
\arguments{
\item{path}{(character) directory in which folders containing final models
were created.}

\item{occ.joint}{(character) the  csv file with training and testing
occurrences combined, or the file containing occurrences used to create final
models; columns must be: species, longitude, latitude, and two or more
columns representing distinct variables.}

\item{occ.ind}{(character) the name of the csv file with independent
occurrences for model evaluation; these occurrences were not used when
creating final models; columns as in \code{occ.joint}. Prepare this
file with \code{\link{prep_independent_swd}}.}

\item{replicates}{(logical) whether or not final models were created
with replicates.}

\item{out.eval}{(character) name of the folder where evaluation results will
be written.}

\item{threshold}{(numeric) the percentage of omission error allowed (E),
default = 5.}

\item{rand.percent}{(numeric) the percentage of data to be used for the
bootstrapping process when calculating partial ROCs; default = 50.}

\item{iterations}{(numeric) the number of times that the bootstrap is going
to be repeated; default = 500.}
}
\value{
A list with two data.frame objects containing results from the
evaluation process, and a folder, in the working directory, containing a
csv file with the results from final model evaluation.
}
\description{
kuenm_feval_swd evaluates final Maxent models in terms of
statistical significance (partial ROC) and omission rates with a user-defined
threshold (E). This function works for models created in SWD format.
}
\details{
This function is used after the creation of final models.
}
