% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_occsplit.R
\name{kuenm_occsplit}
\alias{kuenm_occsplit}
\title{Split occurrence files in training and testing data}
\usage{
kuenm_occsplit(occ.file, train.proportion = 0.5, method = "random",
  name = "occ")
}
\arguments{
\item{occ.file}{(character) name of the csv file with all the occurrences; columns must be:
species, longitude, latitude.}

\item{train.proportion}{(numeric) proportion (from 0 to 1) of data to be used as training
occurrences. The remaining data will be used for testing.}

\item{method}{(character) method for selecting training and testing occurrences. Current
option is "random".}

\item{name}{(character) common name for csv files to be written. A suffix will be added
depending on if the data is the complete set, training set, or testing set of occurrences.}
}
\value{
List with all, training, and testing occurrences. Three csv files will be written in the
working directory according to the name defined in \code{name} plus the suffix _joint
for all records, _train for the training set, and _test for the testing set.
}
\description{
kuenm_occsplit splits occurrences contained in a csv file to obtain training
and testing data based on distinct methods for calibrating models.
}
\examples{
# arguments
occs <- "occurrences.csv"
train_prop <- 0.5
method = "random"
name <- "occ"

# running
data_split <- kuenm_occsplit(occ.file = occs, train.proportion = train_prop,
                             method = method, name = name)
}
