% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_omrat.R
\name{kuenm_omrat}
\alias{kuenm_omrat}
\title{Omission rates calculation for single models}
\usage{
kuenm_omrat(model, threshold = 5, occ.tra, occ.test)
}
\arguments{
\item{model}{a RasterLayer of the model to be evaluated.}

\item{threshold}{(numeric vector) value(s) from 0 to 100 that will be used as thresholds,
default = 5.}

\item{occ.tra}{a numerical matrix containing coordinates of the occurrence data used to create
the ecological niche model to be evaluated; columns must be: longitude and latitude.}

\item{occ.test}{a numerical matrix containing coordinates of the occurrences used to test
the ecological niche model to be evaluated; columns must be: longitude and latitude.}
}
\value{
A named numeric value or numeric vector with the result(s).
}
\description{
kuenm_omrat calculates omission rates of geographic projections
of ecological niche models based on one or multiple user-specified thresholds.
}
\examples{
# single threshold
model <- raster::raster(system.file("extdata/sp_model.tif",
                                       package = "kuenm"))
thres <- 5
data("sp_train", package = "kuenm")
data("sp_test", package = "kuenm")

om_rate <- kuenm_omrat(model, threshold = thres,
                        occ.tra = sp_train, occ.test = sp_test)

# multiple thresholds
thres1 <- c(5, 10, 20)

om_rate <- kuenm_omrat(model, threshold = thres1,
                        occ.tra = sp_train, occ.test = sp_test)
}
