% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_proc.R
\name{kuenm_proc}
\alias{kuenm_proc}
\title{Partial ROC calculation for ecological niche models}
\usage{
kuenm_proc(occ.test, model, threshold = 5, rand.percent = 50,
           iterations = 500, parallel = FALSE)
}
\arguments{
\item{occ.test}{matrix, data.frame, or numeric vector containing coordinates
of occurrences to test model predictions to be evaluated. If matrix or
data.frame, columns must include longitude and latitude in that order.
If numeric, values of suitability in such occurrences. If a matrix or a
data.frame is provided, \code{model} must be a RasterLayer.}

\item{model}{RasterLayer or numeric vector of ecological niche model
predictions to be evaluated. If RasterLayer, layer of predicted suitability.
If numeric vector, predicted suitability values.}

\item{threshold}{(numeric) value from 0 to 100 to represent the percentage of
potential error (E) that the data could have due to any source of uncertainty.
Default = 5.}

\item{rand.percent}{(numeric) percentage of testing data to be used in each
bootstrapped process for calculating the partial ROC. Default = 50.}

\item{iterations}{(numeric) number of bootstrap iterations to be performed;
default = 500.}

\item{parallel}{(logical) argument deprecated. Default = NULL.}
}
\value{
A list with the summary of the results and a data.frame containing
the AUC values and AUC ratios calculated for all iterations.
}
\description{
kuenm_proc applies partial ROC tests to model predictions.
}
\details{
Partial ROC is calculated following Peterson et al. (2008;
\url{http://dx.doi.org/10.1016/j.ecolmodel.2007.11.008}).
}
\examples{
data("sp_test", package = "kuenm")
model <- raster::raster(system.file("extdata/sp_model.tif",
                                       package = "kuenm"))
thres <- 5
rand_perc <- 50
iterac <- 500

p_roc <- kuenm_proc(occ.test = sp_test, model = model, threshold = thres,
                   rand.percent = rand_perc, iterations = iterac)
}
