% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_start.R
\name{kuenm_start}
\alias{kuenm_start}
\title{Creation of an R markdown file for recording all analyses}
\usage{
kuenm_start(file.name)
}
\arguments{
\item{file.name}{(character) is the name of the R markdown file that will be
produced in your working directory.}
}
\value{
An R markdown file with instructions and code for performing all
analyses included in this package.
}
\description{
kuenm_start generates an R markdown file that serves as a guide for
performing all of the analyses included in this package.
}
\examples{
kuenm_start("complete_process")
kuenm_start("bird_modeling_process")
}
