% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_closest.R
\name{kuenm_toclosest}
\alias{kuenm_toclosest}
\title{Move occurrences to closest pixel with environmental data}
\usage{
kuenm_toclosest(data, longitude, latitude, raster.layer, limit.distance)
}
\arguments{
\item{data}{data.frame or matrix of occurrence records. Columns must include
longitude and latitude. Other columns are optional and wont be changed.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{raster.layer}{RasterLayer to be used as a reference.}

\item{limit.distance}{(numeric) maximun distance in km at which an occurrence
could be to be moved. Records farther than this distance wont be moved.}
}
\value{
A data.frame with the corrected coordinates and four additional columns.
The first of the new columns indicates the condition of the coordinates:
Correct, if it was not moved because it was on a pixel with data; Moved, if
it was moved to the nearest pixel; and Not_moved, if it was not moved because
the occurrence was farther than the \code{limit_distance} to the closest pixel.
The second new column indicates the distance to the closest pixel with data.
The other two additional columns will contain the initial longitudes and
latitudes.
}
\description{
kuenm_toclosest helps in changing the longitude and latitude values
of occurrences with no environmental data, so they move to the closest pixel
of a raster layer that contains relevant information. This process prevents
NAs in future analyses.
}
\examples{
data <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                            pattern = "sp_test.csv", full.names = TRUE))

var <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                 pattern = "Mbio_", full.names = TRUE)[1])

raster::plot(var)

out <- data.frame(as.character(data$species[1]), rbind(c(-103, 27), c(-90, 26.5),
                                                       c(-109, 40), c(-70, 41)))
colnames(out) <- colnames(data)

data <- rbind.data.frame(data, out)

points(data[, 2:3])

data1 <- kuenm_toclosest(data, longitude = "Longitude", latitude = "Latitude",
                         raster.layer = var, limit.distance = 200)

points(data1[, 2:3], col = "red")
}
