% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_varcomb.R
\name{kuenm_varcomb}
\alias{kuenm_varcomb}
\title{All potential combinations of a group of variables}
\usage{
kuenm_varcomb(var.dir, out.dir, min.number = 2, in.format = "ascii",
  out.format = "ascii")
}
\arguments{
\item{var.dir}{(character) the name of the folder containing variables that will be combined.}

\item{out.dir}{(character) the name of the folder in which subfolders with distinct combinations of
variables will be written.}

\item{min.number}{(integer) the minimum number of variables per combination. This number must be > 1.
Default = 2.}

\item{in.format}{(character) format of variables in \code{var.dir}. Options are "ascii", "GTiff", and "EHdr" = bil.
Default = "ascii".}

\item{out.format}{(character) format of variables to be written in distinct sets inside \code{out.dir}.
Options are "ascii", "GTiff", and "EHdr" = bil. Default = "ascii".}
}
\value{
A list containing vectors of all the potential combinations of variables. In addition, a folder
named \code{out.dir} with subfolders in which distinct combinations of variables produced are written.
}
\description{
kuenm_varcomb creates multiple sets of variables by grouping them in all their potential combinations.
}
\details{
Sest of variables are written in the working directory and not retained as RasterStacks to avoid
problems related to RAM limitations.

Time of processing will be reduced considerably if \code{in.format} and \code{out.format} coincide
because files will be copied and not loaded and written.
}
\examples{
# This example depends on data stored in your directory
var_dir <- "Variables" # your directory with variables to be combined
out_dir <- "M_variables" # output directory to be created
min_n <- 2
in_format <- "ascii"
out_format <- "GTiff"

comb <- kuenm_varcomb(var.dir = var_dir, out.dir = var_dir, min.number = min_n,
                      in.format = in_format, out.format = out_format)
}
