% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_projchanges.R
\name{model_changes}
\alias{model_changes}
\title{Helper function to calculate model changes}
\usage{
model_changes(calibration.model, current.model, fclim.models,
  result = "continuous", threshold = 5, occ, clim.models, out.dir)
}
\arguments{
\item{calibration.model}{(character) name of calibration model raster name.}

\item{current.model}{(character) name of current model raster name. It can be the same than
\code{calibration.model}.}

\item{fclim.models}{(character) vector of climatic model raster names.}

\item{result}{(character) type of result needed. options are "continuous" and "binary".
Default = "continuous".}

\item{threshold}{(numeric) if \code{result} = "binary", value from 0 to 100 that will be used
as threshold, default = 5.}

\item{occ}{if \code{result} = "binary", a numerical matrix containing coordinates of
the occurrence data used to create the final models; columns must be: longitude and latitude.}

\item{clim.models}{(character) names of that identify climatic models used for project ENMs.
If not defined it is assumed that only one climate model was used.}

\item{out.dir}{(character) name of the folder that will be created to save the binary models
if \code{result} = "binary".}
}
\description{
Helper function to calculate model changes
}
