% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_var_contrib.R
\name{model_var_contrib}
\alias{model_var_contrib}
\title{Evaluation of variable contribution to Maxent final models}
\usage{
model_var_contrib(fmod.dir, model_name = NULL, project, ext.type)
}
\arguments{
\item{fmod.dir}{(character) the  name of the folder in which final models are
(e.g., the output folder after using the \code{\link{kuenm_mod}}) function.
It is important to have only the folders containing the models in this
directory. It can be only one folder or multiple subfolders containing models
for the same species, created with distinct parameter settings. If models were
projected, and the distinct types of extrapolation were used, the name of the
folders contained in this directory should include a pattern describing the
type of extrapolation used (e.g., "EC" for extrapolation and clamping in
Maxent).}

\item{model_name}{(character) pattern to be searched when finding the model of
interest. This pattern does not include the pattern of \code{ext.type}. By
default, NULL, all models are considered.}

\item{project}{(logical) if TRUE, it is assumed that models were projected to
other scenarios (this must be always true if models were produced in SWD
format).}

\item{ext.type}{(character) vector of pattern(s) to be searched in the
folders inside \code{fmod.dir} that identify the extrapolation type(s) of
model projections of interest (e.g., "E", "EC", "NE", or a vector of more
than one of them). Ignored if \code{project} = FALSE.}
}
\value{
A list with results of variable contribution, permutation importance, and
jackknife results. If multiple models are evaluated, a nested list with results
for all models is returned.
}
\description{
model_var_contrib helps to explore variable contribution of
Maxent models created as final models with the functions \code{\link{kuenm_mod}}
or \code{\link{kuenm_mod_swd}}. Variable contribution is measured based on
metrics of contribution percentage, permutation importance, and, if existent,
a jackknife analysis.
}
\details{
When models are created with replicates, the values returned correspond to the
average of such replicates.
}
