% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_mop.R
\name{mop_dist}
\alias{mop_dist}
\title{MOP distance}
\usage{
mop_dist(eudist.matrix, irow, percent)
}
\arguments{
\item{eudist.matrix}{a matrix of Euclidean distance between M and G areas.}

\item{irow}{row from eudist_matrix where the pondarated distance will be compute}

\item{percent}{(numeric) percent of values sampled from te calibration region to calculate the MOP.}
}
\value{
a vector of poderated distances for irow
}
\description{
mop_dist computes a ponderated Euclidean distance for MOP.
}
