% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{mvars}
\alias{mvars}
\title{Raster variables masked to the area where a model is calibrated}
\format{
A RasterStack with 150 rows, 249 columns, 37350 cells, and 4 layers:
\describe{
\item{Temperature}{temperature, in Celsius degrees times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A RasterStack of predictor variables masked to the calibration area where
a model is calibrated. Variables represent four current bioclimatic variables
downloaded from the WorldClim database (\url{http://www.worldclim.org/}).
}
\examples{
mvars <- raster::stack(list.files(system.file("extdata", package = "kuenm"),
                                  pattern = "Mbio_", full.names = TRUE))

summary(mvars)
}
