% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvars_mop.R
\docType{data}
\name{mvars_mop}
\alias{mvars_mop}
\title{Variables of the area in which the ecological niche model is calibrated.}
\format{A RasterStack with 150 rows, 249 columns, 37350 cells, and 4 layers:
\describe{
  \item{Temperature}{temperature, in Celsius degrees times 10.}
  \item{Precipitation}{precipitation, in milimeters.}
  ...
}}
\source{
\url{http://www.ku***/}
}
\usage{
mvars_mop
}
\description{
A dataset containing predictor variables of the area of calibration of the ecological
niche model. Variables represent four current bioclimatic variables dowloaded from the
WorldClim database \url{http://www.worldclim.org/}.
}
\keyword{datasets}
