% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_ceval.R
\name{n.par}
\alias{n.par}
\title{Helper function to calculate the AICc values.}
\usage{
n.par(x)
}
\arguments{
\item{x}{An objects derived from reading the lambdas file created for Maxent.}
}
\description{
Helper function to calculate the AICc values.
}
