% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{n_par}
\alias{n_par}
\title{Helper function to calculate the AICc values (number of parameters).}
\usage{
n_par(x)
}
\arguments{
\item{x}{An object derived from reading the lambdas file created for Maxent.
Use \code{\link[base]{readLines}} function to read the file.}
}
\description{
Helper function to calculate the AICc values (number of parameters).
}
