% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_occsplit.R
\name{occ_randsplit}
\alias{occ_randsplit}
\title{Split occurrences randomly in training and testing data}
\usage{
occ_randsplit(occ, train.proportion = 0.5)
}
\arguments{
\item{occ}{matrix or data.frame with the occurrences to be split. Columns may vary but
species, longitude, and latitue are recommended.}

\item{train.proportion}{(numeric) proportion (from 0 to 1) of data to be used as training
occurrences. The remaining data will be used for testing.}
}
\value{
List with all occurrences (joint), training occurrences (train), and testing (test)
occurrences.
}
\description{
occ_randsplit splits a set of occurrences to obtain training and testing
data randomly.
}
\examples{
# arguments
occs <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                            pattern = "sp_test.csv", full.names = TRUE))
occs <- data.frame(Species = "Species_1", occs)
train_prop <- 0.5

# running
occ_rsplit <- occ_randsplit(occ = occs, train.proportion = train_prop)
}
