% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_analysis_helpers.R
\name{or}
\alias{or}
\title{Omission rates calculation for Maxent SWD predictions}
\usage{
or(prediction, occ.tra, occ.test, threshold = 5)
}
\arguments{
\item{prediction}{matrix of longitude and latidue coordinates, and Maxent
prediction obtained using the SWD format. Prediction coordinates must include
the ones in \code{occ.tra}, and \code{occ.test}.}

\item{occ.tra}{numerical matrix containing coordinates of the occurrence data
used to create the prediction to be evaluated; columns must be: longitude and
latitude.}

\item{occ.test}{numerical matrix containing coordinates of the occurrences
used to test the prediction to be evaluated; columns must be: longitude and
latitude.}

\item{threshold}{(numeric) vector of value(s) from 0 to 100 that will be used
as thresholds, default = 5.}
}
\value{
A named numeric value or numeric vector with the result(s).
}
\description{
or calculates omission rates of numerical projections of ecological
niche models based on one or multiple user-specified thresholds.
}
