% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_var_contrib.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Helper to plot variable contribution to single models}
\usage{
plot_contribution(contribution_list)
}
\arguments{
\item{contribution_list}{a list of results obtained with
\code{\link{explore_var_contrib}}.}
}
\description{
Helper to plot variable contribution to single models
}
