% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{plot_proc_aicc}
\alias{plot_proc_aicc}
\title{Helper function to plot omission rate and AICc results}
\usage{
plot_proc_aicc(summary.calibration)
}
\arguments{
\item{summary.calibration}{data.frame containing the summary of all metrics
calculated for all models during calibration.}
}
\description{
Helper function to plot omission rate and AICc results
}
