% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_swd.R
\name{prep_independent_swd}
\alias{prep_independent_swd}
\title{Helper to prepare independent occurrences when using SWD format}
\usage{
prep_independent_swd(
  occ,
  species,
  longitude,
  latitude,
  raster.layers,
  save = FALSE,
  name.occ
)
}
\arguments{
\item{occ}{data.frame containing occurrence records of the species of interest.
Mandatory columns are: species, longitude, and latitude. Other columns will
be ignored.}

\item{species}{(character) name of column containing species name.}

\item{longitude}{(character) name of column containing longitude values.}

\item{latitude}{(character) name of column containing latitude values.}

\item{raster.layers}{RasterStack of predictor variables masked to the area
where the model was calibrated.}

\item{save}{(logical) whether or not to a write csv file containing
independent occurrences prepared to be used in model evaluation. The file
will contain additional columns with the values of the variables for each
coordinate. Default = FALSE.}

\item{name.occ}{(character) name to be used for the file with occurrence
records to be written (e.g., "independent_occ").}
}
\value{
A data.frame with the prepared independent occurrences.
}
\description{
Helper to prepare independent occurrences when using SWD format
}
