% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_swd.R
\name{prepare_swd}
\alias{prepare_swd}
\title{Prepare data for SWD maxent calibration processes}
\usage{
prepare_swd(occ, species, longitude, latitude,
            data.split.method = "random", train.proportion = 0.5,
            raster.layers, sample.size = 10000, save = FALSE,
            name.occ, name.back, set.seed = 1)
}
\arguments{
\item{occ}{data.frame containing occurrence records of the species of interest.
Mandatory columns are: species, longitude, and latitude. Other columns will
be ignored.}

\item{species}{(character) name of column containing species name.}

\item{longitude}{(character) name of column containing longitude values.}

\item{latitude}{(character) name of column containing latitude values.}

\item{data.split.method}{(character) name of the method to split training and
testing records. Default and only option for now = "random".}

\item{train.proportion}{(numeric) proportion of records to be used for training
models. Default = 0.5}

\item{raster.layers}{RasterStack of predictor variables masked to the area
where the model will be calibrated.}

\item{sample.size}{(numeric) number of points to represent the background for
the model. Default = 10000}

\item{save}{(logical) whether or not to write csv files containing all, train,
and test occurrences, as well as the background. All files will contain
additional columns with the values of the variables for each coordinate.
Default = FALSE}

\item{name.occ}{(character) name to be used for files with occurrence records.
Only one name is needed, a sufix will be added to represent all (_join),
_train, and _test records (e.g., "occurrences").}

\item{name.back}{name for the csv file containing background coordinates
(e.g., "background").}

\item{set.seed}{seed to be used when sampling background and splitting records.
Default = 1}
}
\description{
prepare_swd helps to create csv files containing occurrence
records (all, train, and test records) and background coordinates, together
with values of predictor variables, that later can be used to run model
calibration in Maxent using the SWD format.
}
\examples{
# data
occ <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                           pattern = "sp_joint.csv", full.names = TRUE))
occ <- data.frame(Species = "A_americanum", occ)

mvars <- raster::stack(list.files(system.file("extdata", package = "kuenm"),
                                  pattern = "Mbio_", full.names = TRUE))

# preparing swd
prep <- prepare_swd(occ, species = "Species", longitude = "Longitude",
                    latitude = "Latitude", raster.layers = mvars,
                    sample.size = 5000)
}
