% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_analysis_helpers.R
\name{proc_or_aicc}
\alias{proc_or_aicc}
\title{Partial ROC, omission rates, and AICc calculations in concert (helper)}
\usage{
proc_or_aicc(occ.joint, occ.tra, occ.test, raw.folders, log.folders,
             threshold = 5, rand.percent = 50, iterations = 500, kept = TRUE)
}
\arguments{
\item{occ.joint}{(character) the name of csv file with training and testing
occurrences combined; columns must be: species, longitude, and latitude.}

\item{occ.tra}{(character) the name of the csv file with the training
occurrences; columns as in \code{occ.joint}.}

\item{occ.test}{(character) the name of the csv file with the evaluation
occurrences; columns as in \code{occ.joint}.}

\item{raw.folders}{(character) vector of names of directories containing
models created with all occurrences and raw outputs.}

\item{log.folders}{(character) vector of names of directories containing
models created with training occurrences and logistic outputs.}

\item{threshold}{(numeric) the percentage of training data omission error
allowed (E); default = 5.}

\item{rand.percent}{(numeric) the percentage of data to be used for the
bootstraping process when calculating partial ROCs; default = 50.}

\item{iterations}{(numeric) the number of times that the bootstrap is going
to be repeated; default = 500.}

\item{kept}{(logical) if FALSE, all candidate models will be erased after
evaluation, default = TRUE.}
}
\value{
A data.frame with the results of partial ROC, omission rates, and AICc metrics
for all candidate models.
}
\description{
proc_or_aicc performs a series of step by step processes that
help to read files from directores, extract necessary data, and evaluate
Maxent predictions based on partial ROC, omission rates, and AICc values.
}
