% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{run_maxent}
\alias{run_maxent}
\title{Helper function to run maxent.jar from R}
\usage{
run_maxent(batch, maxent.path, add_path = TRUE, wait = FALSE)
}
\arguments{
\item{batch}{(character) name of the batch file (bash for Unix) with the code
to create all candidate models.}

\item{maxent.path}{(character) the path were the maxent.jar file is in your
computer.}

\item{add_path}{(logical) whether to add full path to \code{batch}.}

\item{wait}{(logical) whether R waits until the running is done or not.}
}
\description{
Helper function to run maxent.jar from R
}
