% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{sp_joint}
\alias{sp_joint}
\title{A set of occurrence records for ecological niche models}
\format{
A data frame with 178 rows and 2 columns.
\describe{
\item{Longitude}{longitude, in decimal degrees.}
\item{Latitude}{latitude, in decimal degrees.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A data.frame containing occurrence records of a tick (\emph{Amblyomma americanum})
across North America. The data combines records for training and testing.
}
\examples{
data("sp_joint", package = "kuenm")

head(sp_joint)
}
