% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_mod.R
\docType{data}
\name{sp_mod}
\alias{sp_mod}
\title{An ecological niche model created with Maxent}
\format{A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability, in probability values.}
}}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\usage{
sp_mod
}
\description{
A RasterLayer containing an ecological niche model for the a tick (\emph{Amblyomma americanum})
that was created as part of the candidate models during the calibration process.
}
\examples{
model <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                   pattern = "sp_model.tif", full.names = TRUE))

raster::plot(model)
}
\keyword{datasets}
