% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_mod_joint.R
\name{sp_mod_joint}
\alias{sp_mod_joint}
\title{An ecological niche model created with Maxent with raw output}
\format{
A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability, in probability values.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A RasterLayer containing an ecological niche model for the a tick (\emph{Amblyomma americanum})
that was created with all occurrences and raw output.
}
\examples{
model <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                   pattern = "sp_model_joint.tif", full.names = TRUE))

raster::plot(model)
}
