% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{sp_model}
\alias{sp_model}
\title{A raster output of an ecological niche model created with Maxent (logistic)}
\format{
A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability values.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A RasterLayer containing an ecological niche model for the tick
(\emph{Amblyomma americanum}) that was created as part of the candidate models
during a calibration process.
}
\examples{
sp_model <- raster::raster(system.file("extdata/sp_model.tif",
                                       package = "kuenm"))

summary(sp_model)
}
