% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_test.R
\docType{data}
\name{sp_test}
\alias{sp_test}
\title{A set of occurrence records for testing candidate ecological niche models.}
\format{A data frame with 89 rows and 2 columns.
\describe{
  \item{Longitude}{longitude, in decimal degrees.}
  \item{Latitude}{latitude, in decimal degrees.}
  ...
}}
\source{
\url{http://www.ku***/}
}
\usage{
sp_test
}
\description{
A dataset conatining occurrence recods of the Turkey tick in North America, that
are used to test candidate ecological niche models during calibration.
}
\keyword{datasets}
