% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_analysis_helpers.R
\name{summary_calibration}
\alias{summary_calibration}
\title{Helper to summarize all results from model calibration exercises}
\usage{
summary_calibration(proc.or.aicc.results, selection = "OR_AICc")
}
\arguments{
\item{proc.or.aicc.results}{data.frame with results from evaluation of all
candidate models. Generally the output of \code{\link{proc_or_aicc}}.}

\item{selection}{(character) model selection criterion, can be "OR_AICc",
"AICc", or "OR"; OR = omission rates. Default = "OR_AICc", which means that
among models that are statistically significant and that present omission
rates below the threshold, those with delta AICc up to 2 will be selected.
See details for other selection criteria.}
}
\value{
A list with all results that need to be written to produce the evaluation report.
}
\description{
Helper to summarize all results from model calibration exercises
}
\details{
Other selecton criteria are described below: If "AICc" criterion is chosen,
all significant models with delta AICc up to 2 will be selected If "OR" is
chosen, the 10 first significant models with the lowest omission rates will
be selected.
}
