% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_closest.R
\name{to_closest}
\alias{to_closest}
\title{Move occurrences to closest pixel with environmental data}
\usage{
to_closest(data, longitude, latitude, raster.layer)
}
\arguments{
\item{data}{data.frame or matrix of occurrence records. Columns must include
longitude and latitude. Other columns are optional and wont be changed.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{raster.layer}{RasterLayer to be used as a reference.}
}
\value{
data.frame or matrix with the corrected coordinates.
}
\description{
to_closest helps in changing the longitude and latitude values
of occurrences with no environmental data, so they move to the closest pixel
of a raster layer that contains relevant information. This process prevents
NAs in future analyses.
}
\examples{
data <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                            pattern = "sp_test.csv", full.names = TRUE))

var <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                 pattern = "Mbio_", full.names = TRUE)[1])

raster::plot(var)

out <- data.frame(as.character(data$species[1]), rbind(c(-103, 27), c(-90, 26.5),
                                                       c(-109, 40), c(-70, 41)))
colnames(out) <- colnames(data)

data <- rbind.data.frame(data, out)

points(data[, 2:3])

data1 <- to_closest(data, longitude = "Longitude", latitude = "Latitude",
                    raster.layer = var)

points(data1[, 2:3], col = "red")
}
