% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{wait_written_done}
\alias{wait_written_done}
\title{Helper function to wait until a file writing is done}
\usage{
wait_written_done(file)
}
\arguments{
\item{file}{(character) name of the file of interest.}
}
\description{
Helper function to wait until a file writing is done
}
