% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_functions.R
\name{a_pred_pres}
\alias{a_pred_pres}
\title{Helper function to compute the area (number of pixels) that a certain threshold has.}
\usage{
a_pred_pres(inrast)
}
\arguments{
\item{inrast}{A raster class object of a continuous model output.}
}
\description{
Helper function to compute the area (number of pixels) that a certain threshold has.
}
