% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierpart_analyses.R
\name{hierpart_analyses}
\alias{hierpart_analyses}
\title{Performing hierarchical partitioning analyses}
\usage{
hierpart_analyses(tables.folder, out.dir, kept = FALSE)
}
\arguments{
\item{tables.folder}{(character) name of the folder where the tables created with the
function {\link{hierpart_tables}} are.}

\item{out.dir}{(character) name of the output directory where results will be written.}
}
\value{
All the results from performing hierarchical partitioning analyses of the
varianze in ecological niche models.
}
\description{
Helper function to perform hierarchical partitioning analyses
of the varianze comming from distinct sources in ecological niche models.
}
