% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWD_short_helpers.R
\name{html_calibration}
\alias{html_calibration}
\title{Helper function to create an HTML file summarizing results}
\usage{
html_calibration(path = getwd(), file.name)
}
\arguments{
\item{path}{directory where the HTML file will be written; default = current.}

\item{file.name}{(character) name of the HTML file.}
}
\value{
HTML file in \code{path}
}
\description{
Helper function to create an HTML file summarizing results
}
