% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ku.enm.cal.R
\name{ku.enm.cal}
\alias{ku.enm.cal}
\title{Creation of candidate models for calibration}
\usage{
ku.enm.cal(occ.joint, occ.tra, M.var.dir, batch, out.dir, reg.mult,
  f.clas = "all", run = TRUE)
}
\arguments{
\item{occ.joint}{(character) is the name of the csv file with all the occurrences; columns must be: species, longitude, latitude.}

\item{occ.tra}{(character) is the name of the csv file with the calibration occurrences; columns equal to occ.joint.}

\item{M.var.dir}{(character) is the name of the folder containing other folders with different sets of environmental variables.}

\item{batch}{(character) name of the batch file with the code to create all candidate models.}

\item{out.dir}{(character) name of the folder that will contain all calibration model subfolders.}

\item{reg.mult}{(numeric vector) regularization multiplier(s) to be evaluated.}

\item{f.clas}{(character) feature clases can be selected from  five different combination sets or manually.
Combination sets are: "all", "basic", "no.t.h", "no.h", and "no.t". Default = "all".
basic = "l", "lq", "lqp", "lqpt", "lqpth". Combinations "no.t.h", "no.h", and "no.t", exclude t and/or h.
Options from the following list can be selected manually:
"l", "q", "p", "t", "h", "lq", "lp", "lt", "lh", "qp", "qt", "qh",
"pt", "ph", "th", "lqp", "lqt", "lqh", "lpt", "lph", "qpt", "qph",
"qth", "pth", "lqpt", "lqph", "lqth", "lpth", and "lqpth".}

\item{run}{(logical) if true the batch runs after its creation, if false it will only be created and its runnig would be
manual, default = TRUE.}
}
\value{
A folder named out.dir with all the subfolders to save Maxent results when running the .bat file.
A .bat file containing the java codes to run the calibration models, it will run auotmatically or on some
computers a dialog box will ask if running is allowed.
}
\description{
ku.enm.cal creates and executes a batch file for generating candidate models in Maxent
to test multiple parameter combinations, including distinct regularization multiplier values,
various feature classes, and different sets of environmental variables.
}
\details{
Java needs to be installed in the computer and maxent.jar needs to be located in the working directory.
Java can be obtained from \url{https://java.com/es/download/manual.jsp}. Maxent can be downloaded from
\url{https://biodiversityinformatics.amnh.org/open_source/maxent/}
}
