% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_aicc.R
\name{kuenm_aicc}
\alias{kuenm_aicc}
\title{AICc calculation of Maxent models.}
\usage{
kuenm_aicc(occ, model, npar)
}
\arguments{
\item{occ}{a numerical matrix containing coordinates of the occurrences used to create
the ecological niche models to be evaluated; columns must be: longitude and latitude.}

\item{model}{a RasterLayer or RasterStack of ecological niche models created using Maxent
with the Raw output.}

\item{npar}{(numeric) vector of number of parameters for \code{model}. Lenght must
correspond with number of models to be evaluated. Use function \code{\link{n.par}} to obtain
number of parameters for each model.}
}
\value{
A dataframe containing values of AICc, delta AICc, weight of AICc, and number of parameters.
Number of rows of this dataframe correspond to number of models evaluated.
}
\description{
kuenm_aicc calculates the Akaike information criterion corrected for small
sample sizes (AICc) to single or multiple models produced with Maxent.
}
\details{
This function is a modificatio of the \code{\link[ENMeval]{calc.aicc}} from the ENMeval package.
Changes help to make calcutions faster because of a better management of raster values (especially
when calculations are performed for multiple models).
}
\examples{
occ <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                           pattern = "sp_joint.csv", full.names = TRUE))
model <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                   pattern = "sp_model_joint.tif", full.names = TRUE))

lbds <- readLines(list.files(system.file("extdata", package = "kuenm"), # lambdas file
                                         pattern = "lambdas_model_joint.lambdas", full.names = TRUE))
npar <- n.par(lbds) # counting number of parameters

aicc <- kuenm_aicc(occ = occ, model = model, npar = npar)
}
