% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm_feval.R
\name{kuenm_feval}
\alias{kuenm_feval}
\title{Evaluation of final Maxent models with independent data}
\usage{
kuenm_feval(path, occ.joint, occ.ind, replicates, out.eval, threshold = 5,
  rand.percent = 50, iterations = 500)
}
\arguments{
\item{path}{(character) directory in which folders containig final models are being created or
were created.}

\item{occ.joint}{(character) the  csv file with training and testing occurrences combined,
or the file containing occurrences used to create final models; columns must be: species,
longitude, latitude.}

\item{occ.ind}{(character) the name of the csv file with independent occurrences for model
evaluation; these occurrences were not used when creating final models; columns as in occ.joint.}

\item{replicates}{(logical) whether or not final models were created performing replicates.}

\item{out.eval}{(character) name of the folder where evaluation results will be written.}

\item{threshold}{(numeric) the percentage of omission error allowed (E), default = 5.}

\item{rand.percent}{(numeric) the percentage of data to be used for the bootstrapping process
when calculating partial ROCs; default = 50.}

\item{iterations}{(numeric) the number of times that the bootstrap is going to be repeated;
default = 500.}
}
\value{
A list with two dataframes containing results from the evaluation process, and
a folder, in the working directory, containing a csv file with the final models evaluation
results.
}
\description{
kuenm_feval evaluates final Maxent models in terms of statistical
significance (partial ROC) and omission rates with a user-defined threshold (E).
}
\details{
This function is used after or during the creation of final models.
}
