% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvars_mop.R
\docType{data}
\name{mvars_mop}
\alias{mvars_mop}
\title{Variables of the area in which the ecological niche model is calibrated.}
\format{
A RasterStack with 150 rows, 249 columns, 37350 cells, and 4 layers:
\describe{
\item{Temperature}{temperature, in Celsius degrees times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\usage{
mvars_mop
}
\description{
A dataset containing predictor variables for the calibration area for the ecological
niche model. Variables represent four current bioclimatic variables dowloaded from the
WorldClim database (\url{http://www.worldclim.org/}).
}
\examples{
mvars <- raster::stack(list.files(system.file("extdata", package = "kuenm"),
                                  pattern = "Mbio_", full.names = TRUE))

raster::plot(mvars[[1]])
}
\keyword{datasets}
