% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_out.R
\name{plot_out}
\alias{plot_out}
\title{Detection of environmental values ouside the calibration area of a model}
\usage{
plot_out(M1, G1)
}
\arguments{
\item{M1}{a numeric matrix containing values of all environmental variables in the calibration area.}

\item{G1}{a numeric matrix containing values of all environmental variables in the full area of interest.}
}
\description{
plot.out for calculating a mobility-oriented parity layer.
This function is designed to be used specifically in the \code{\link{kuenm_mop}} function.
}
