% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lambdas.R
\name{sp_lambdas}
\alias{sp_lambdas}
\title{A lambdas file resulted from a modeling process in Maxent}
\format{A lambdas file.
\describe{
\item{parameters}{number of parameters in the Maxent model.}
}}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A lambdas file resulted from a model created in Maxent with raw output for a tick
(\emph{Amblyomma americanum}) in North America. This file is used to calculate number
of parameters in the model, which is needed while calculating AICc values.
}
\examples{
lbds <- readLines(list.files(system.file("extdata", package = "kuenm"), # lambdas file
                                          pattern = "lambdas_model_joint.lambdas", full.names = TRUE))

head(lbds)
}
