% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_mod.R
\docType{data}
\name{sp_mod}
\alias{sp_mod}
\title{An ecological niche model created with Maxent.}
\format{A RasterLtack with 150 rows, 249 columns, and 37350 cells:
\describe{
  \item{Suitability}{suitability, in probability values.}
  ...
}}
\source{
\url{http://www.ku***/}
}
\usage{
sp_mod
}
\description{
A raster layer containing an ecological niche model for the Turkey tick that was
created as part of the candidate models for teh calibration process.
}
\keyword{datasets}
