% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/sp_mod_joint.R
\name{sp_mod_joint}
\alias{sp_mod_joint}
\title{A raster output of an ecological niche model created with Maxent (raw)}
\format{
A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability values.}
}

A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability, in probability values.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}

\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\description{
A RasterLayer containing an ecological niche model for the a tick
(\emph{Amblyomma americanum}) that was created with all occurrences.

A RasterLayer containing an ecological niche model for the a tick (\emph{Amblyomma americanum})
that was created with all occurrences and raw output.
}
\examples{
sp_model_joint <- raster::raster(system.file("extdata/sp_model_joint.tif",
                                             package = "kuenm"))

summary(sp_model_joint)
model <- raster::raster(list.files(system.file("extdata", package = "kuenm"),
                                   pattern = "sp_model_joint.tif", full.names = TRUE))

raster::plot(model)
}
