% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_test.R
\docType{data}
\name{sp_test}
\alias{sp_test}
\title{A set of occurrence records for testing candidate ecological niche models.}
\format{A data frame with 89 rows and 2 columns.
\describe{
  \item{Longitude}{longitude, in decimal degrees.}
  \item{Latitude}{latitude, in decimal degrees.}
}}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\usage{
sp_test
}
\description{
A dataset containing occurrence recods of the Turkey Tick (*Amblyomma americanum*)
in North America, used to test candidate ecological niche models during calibration.
}
\keyword{datasets}
