% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_train.R
\docType{data}
\name{sp_train}
\alias{sp_train}
\title{A set of occurrence records for training candidate ecological niche models.}
\format{A data frame with 89 rows and 2 columns.
\describe{
  \item{Longitude}{longitude, in decimal degrees.}
  \item{Latitude}{latitude, in decimal degrees.}
  ...
}}
\source{
\url{http://www.ku***/}
}
\usage{
sp_train
}
\description{
A dataset conatining occurrence recods of the Turkey tick in North America, that
are used to train candidate ecological niche models during calibration.
}
\keyword{datasets}
