% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_train.R
\docType{data}
\name{sp_train}
\alias{sp_train}
\title{A set of occurrence records for training candidate ecological niche models}
\format{A data frame with 89 rows and 2 columns.
\describe{
\item{Longitude}{longitude, in decimal degrees.}
\item{Latitude}{latitude, in decimal degrees.}
}}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\usage{
sp_train
}
\description{
A dataset containing occurrence recods of a tick (\emph{Amblyomma americanum}) across
North America, used to train candidate ecological niche models during calibration.
}
\examples{
occtr <- read.csv(list.files(system.file("extdata", package = "kuenm"),
                            pattern = "sp_train.csv", full.names = TRUE))

head(occtr)
}
\keyword{datasets}
